rename: Store 위치 변경
